pose_viewer/requirements.txt
openvino
opencv-python
pillow
pyyaml
numpy
